+++ src/revolut/resources/__init__.py
@@ -2,6 +2,7 @@
 
 from .customers import AsyncCustomersResource, CustomersResource
 from .orders import AsyncOrdersResource, OrdersResource
+from .payment_methods import AsyncPaymentMethodsResource, PaymentMethodsResource
 from .payments import AsyncPaymentsResource, PaymentsResource
 from .refunds import AsyncRefundsResource, RefundsResource
 from .webhooks import AsyncWebhooksResource, WebhooksResource
@@ -9,11 +10,13 @@ from .webhooks import AsyncWebhooksResource, WebhooksResource
 __all__ = [
     "AsyncCustomersResource",
     "AsyncOrdersResource",
+    "AsyncPaymentMethodsResource",
     "AsyncPaymentsResource",
     "AsyncRefundsResource",
     "AsyncWebhooksResource",
     "CustomersResource",
     "OrdersResource",
+    "PaymentMethodsResource",
     "PaymentsResource",
     "RefundsResource",
     "WebhooksResource",